+++ packages/replay-internal/test/integration/autoSaveSession.test.ts
@@ -2,7 +2,7 @@
  * @vitest-environment jsdom
  */
 
-import '../utils/use-fake-timers';
+import '../utils/mock-internal-setTimeout';
 import { EventType } from '@sentry-internal/rrweb';
 import { afterEach, beforeAll, describe, expect, test, vi } from 'vitest';
 import { saveSession } from '../../src/session/saveSession';
