+++ packages/replay-internal/test/integration/coreHandlers/handleGlobalEvent.test.ts
@@ -2,7 +2,7 @@
  * @vitest-environment jsdom
  */
 
-import '../../utils/use-fake-timers';
+import '../../utils/mock-internal-setTimeout';
 import type { Event } from '@sentry/core';
 import { getClient } from '@sentry/core';
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
