+++ packages/replay-internal/test/unit/coreHandlers/handleClick.test.ts
@@ -2,7 +2,7 @@
  * @vitest-environment jsdom
  */
 
-import '../../utils/use-fake-timers';
+import '../../utils/mock-internal-setTimeout';
 import type { Breadcrumb } from '@sentry/core';
 import { beforeAll, beforeEach, describe, expect, it, test, vi } from 'vitest';
 import { ClickDetector, ignoreElement } from '../../../src/coreHandlers/handleClick';
