+++ packages/replay-internal/test/unit/coreHandlers/handleNetworkBreadcrumbs.test.ts
@@ -2,7 +2,7 @@
  * @vitest-environment jsdom
  */
 
-import '../../utils/use-fake-timers';
+import '../../utils/mock-internal-setTimeout';
 import type {
   Breadcrumb,
   BreadcrumbHint,
