+++ packages/replay-internal/test/unit/coreHandlers/util/addBreadcrumbEvent.test.ts
@@ -2,7 +2,7 @@
  * @vitest-environment jsdom
  */
 
-import '../../../utils/use-fake-timers';
+import '../../../utils/mock-internal-setTimeout';
 import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import { addBreadcrumbEvent } from '../../../../src/coreHandlers/util/addBreadcrumbEvent';
 import type { EventBufferArray } from '../../../../src/eventBuffer/EventBufferArray';
