+++ packages/replay-internal/test/unit/util/debounce.test.ts
@@ -1,4 +1,4 @@
-import '../../utils/use-fake-timers';
+import '../../utils/mock-internal-setTimeout';
 import { beforeAll, describe, expect, it, vi } from 'vitest';
 import { debounce } from '../../../src/util/debounce';
